refactor: position item
