--- apps/web/src/views/universalFarms/components/PositionItem.tsx
@@ -1,723 +1,0 @@
-import { chainNames } from '@pancakeswap/chains'
-import { Protocol } from '@pancakeswap/farms'
-import { useTheme } from '@pancakeswap/hooks'
-import { useTranslation } from '@pancakeswap/localization'
-import { ERC20Token } from '@pancakeswap/sdk'
-import { Currency, CurrencyAmount, Price, Token } from '@pancakeswap/swap-sdk-core'
-import { unwrappedToken } from '@pancakeswap/tokens'
-import {
-  AddIcon,
-  Box,
-  Button,
-  Column,
-  FeeTier,
-  Flex,
-  IconButton,
-  MinusIcon,
-  Row,
-  Skeleton,
-  SwapHorizIcon,
-  Tag,
-  Text,
-  useMatchBreakpoints,
-  useModalV2,
-} from '@pancakeswap/uikit'
-import { formatBigInt } from '@pancakeswap/utils/formatBalance'
-import { Bound, DoubleCurrencyLogo, FiatNumberDisplay } from '@pancakeswap/widgets-internal'
-import BigNumber from 'bignumber.js'
-import { RangeTag } from 'components/RangeTag'
-import { TokenPairImage } from 'components/TokenImage'
-import { useCakePrice } from 'hooks/useCakePrice'
-import { useCurrencyUsdPrice } from 'hooks/useCurrencyUsdPrice'
-import { formatTickPrice } from 'hooks/v3/utils/formatTickPrice'
-import NextLink from 'next/link'
-import { memo, useCallback, useMemo, useState } from 'react'
-import { useStakedPositionsByUser } from 'state/farmsV3/hooks'
-import { getPoolAddressByToken, useExtraV3PositionInfo, usePoolInfo } from 'state/farmsV4/hooks'
-import type { PositionDetail, StableLPDetail, V2LPDetail } from 'state/farmsV4/state/accountPositions/type'
-import { type PoolInfo } from 'state/farmsV4/state/type'
-import styled from 'styled-components'
-import currencyId from 'utils/currencyId'
-import { logGTMClickStakeFarmEvent } from 'utils/customGTMEventTracking'
-import useFarmV3Actions from 'views/Farms/hooks/v3/useFarmV3Actions'
-import { v2Fee } from 'views/PoolDetail/hooks/useStablePoolFee'
-import { useCheckShouldSwitchNetwork } from '../hooks'
-import { addQueryToPath } from '../utils'
-import { PoolApyButton } from './PoolApyButton'
-import { StakeModal } from './StakeModal'
-import { StyledNextLink } from './StyledCard'
-
-const Container = styled(Flex)`
-  padding: 16px;
-  align-items: flex-start;
-  position: relative;
-  gap: 12px;
-  border-radius: 20px;
-  border: 1px solid ${({ theme }) => theme.colors.cardBorder};
-  border-bottom-width: 2px;
-  background: ${({ theme }) => theme.card.background};
-  margin: 8px 0;
-`
-
-const DetailsContainer = styled(Flex)`
-  flex-direction: column;
-  justify-content: space-between;
-  flex: 1;
-  gap: 8px;
-  color: ${({ theme }) => theme.colors.textSubtle};
-
-  ${({ theme }) => theme.mediaQueries.md} {
-    flex-direction: row;
-  }
-`
-
-const TagCell = styled(Flex)`
-  position: absolute;
-  right: 0;
-  top: 0;
-  padding: 16px;
-`
-
-interface IPriceRangeProps {
-  currency0: Currency
-  currency1: Currency
-  priceUpper?: Price<Token, Token>
-  priceLower?: Price<Token, Token>
-  tickAtLimit: {
-    LOWER?: boolean
-    UPPER?: boolean
-  }
-}
-
-const useTotalPriceUSD = ({
-  currency0,
-  currency1,
-  amount0,
-  amount1,
-}: {
-  currency0: Currency | null | undefined
-  currency1: Currency | null | undefined
-  amount0?: CurrencyAmount<ERC20Token>
-  amount1?: CurrencyAmount<ERC20Token>
-}) => {
-  const { data: currency0PriceFromApi } = useCurrencyUsdPrice(currency0, {
-    enabled: !!currency0,
-  })
-  const { data: currency1PriceFromApi } = useCurrencyUsdPrice(currency1, {
-    enabled: !!currency1,
-  })
-  return useMemo(() => {
-    return (
-      Number(currency0PriceFromApi) * Number(amount0 ? amount0.toExact() : 0) +
-      Number(currency1PriceFromApi) * Number(amount1 ? amount1.toExact() : 0)
-    )
-  }, [amount0, amount1, currency0PriceFromApi, currency1PriceFromApi])
-}
-
-const PriceRange = memo(({ currency1, currency0, priceLower, priceUpper, tickAtLimit }: IPriceRangeProps) => {
-  const [priceBaseInvert, setPriceBaseInvert] = useState(false)
-  const {
-    t,
-    currentLanguage: { locale },
-  } = useTranslation()
-
-  const toggleSwitch: React.MouseEventHandler<HTMLOrSVGElement> = useCallback(
-    (e) => {
-      e.preventDefault()
-      setPriceBaseInvert(!priceBaseInvert)
-    },
-    [priceBaseInvert],
-  )
-
-  return priceUpper && priceLower ? (
-    <>
-      {t('Min %minAmount%', {
-        minAmount: formatTickPrice(
-          priceBaseInvert ? priceUpper.invert() : priceLower,
-          tickAtLimit,
-          Bound.LOWER,
-          locale,
-        ),
-      })}{' '}
-      /{' '}
-      {t('Max %maxAmount%', {
-        maxAmount: formatTickPrice(
-          priceBaseInvert ? priceLower.invert() : priceUpper,
-          tickAtLimit,
-          Bound.UPPER,
-          locale,
-        ),
-      })}{' '}
-      {t('of %assetA% per %assetB%', {
-        assetA: priceBaseInvert ? currency1.symbol : currency0.symbol,
-        assetB: priceBaseInvert ? currency0.symbol : currency1.symbol,
-      })}
-      <IconButton onClick={toggleSwitch} variant="text" scale="xs">
-        <SwapHorizIcon color="textSubtle" ml="2px" />
-      </IconButton>
-    </>
-  ) : null
-})
-
-export const PositionV2Item = memo(
-  ({ data, detailMode }: { data: V2LPDetail; pool?: PoolInfo; detailMode?: boolean }) => {
-    const { pair, deposited0, deposited1 } = data
-
-    const unwrappedToken0 = unwrappedToken(pair.token0)
-    const unwrappedToken1 = unwrappedToken(pair.token1)
-    const totalPriceUSD = useTotalPriceUSD({
-      currency0: unwrappedToken0,
-      currency1: unwrappedToken1,
-      amount0: deposited0,
-      amount1: deposited1,
-    })
-
-    const feeAmount = useMemo(() => Number(v2Fee.multiply(100).toFixed(2)), [])
-    const pool = usePoolInfo({ poolAddress: data.pair.liquidityToken.address, chainId: pair.chainId })
-
-    return (
-      <PositionItemDetail
-        chainId={pair.chainId}
-        pool={pool}
-        link={`/v2/pair/${currencyId(unwrappedToken0)}/${currencyId(unwrappedToken1)}`}
-        totalPriceUSD={totalPriceUSD}
-        currency0={unwrappedToken0}
-        currency1={unwrappedToken1}
-        removed={false}
-        outOfRange={false}
-        protocol={data.protocol}
-        fee={feeAmount}
-        amount0={deposited0}
-        amount1={deposited1}
-        detailMode={detailMode}
-      />
-    )
-  },
-)
-
-export const PositionStableItem = memo(
-  ({ data, detailMode }: { data: StableLPDetail; pool?: PoolInfo; detailMode?: boolean }) => {
-    const { pair, deposited0, deposited1 } = data
-
-    const totalPriceUSD = useTotalPriceUSD({
-      currency0: pair.token0,
-      currency1: pair.token1,
-      amount0: deposited0,
-      amount1: deposited1,
-    })
-    const pool = usePoolInfo({ poolAddress: pair.stableSwapAddress, chainId: pair.liquidityToken.chainId })
-
-    return (
-      <PositionItemDetail
-        chainId={pair.liquidityToken.chainId}
-        pool={pool}
-        link={`/stable/${pair.lpAddress}`}
-        totalPriceUSD={totalPriceUSD}
-        currency0={pair.token0}
-        currency1={pair.token1}
-        removed={false}
-        outOfRange={false}
-        protocol={data.protocol}
-        fee={Number(pair.fee.numerator)}
-        amount0={deposited0}
-        amount1={deposited1}
-        detailMode={detailMode}
-      />
-    )
-  },
-)
-interface IPositionV3ItemProps {
-  data: PositionDetail
-  poolInfo?: PoolInfo
-  detailMode?: boolean
-}
-
-export const PositionV3Item = memo(({ data, detailMode }: IPositionV3ItemProps) => {
-  const { currency0, currency1, removed, outOfRange, priceUpper, priceLower, tickAtLimit, position } =
-    useExtraV3PositionInfo(data)
-
-  const poolAddress = getPoolAddressByToken(data.chainId, data.token0, data.token1, data.fee)
-  const pool = usePoolInfo({ poolAddress, chainId: data.chainId })
-
-  const totalPriceUSD = useTotalPriceUSD({
-    currency0,
-    currency1,
-    amount0: position?.amount0,
-    amount1: position?.amount1,
-  })
-
-  const desc =
-    currency1 && currency0 ? (
-      <PriceRange
-        currency0={currency0}
-        currency1={currency1}
-        priceLower={priceLower}
-        priceUpper={priceUpper}
-        tickAtLimit={tickAtLimit}
-      />
-    ) : null
-
-  return (
-    <PositionItemDetail
-      chainId={data.chainId}
-      link={`/liquidity/${data.tokenId}`}
-      pool={pool}
-      totalPriceUSD={totalPriceUSD}
-      amount0={position?.amount0}
-      amount1={position?.amount1}
-      desc={desc}
-      currency0={currency0}
-      currency1={currency1}
-      removed={removed}
-      outOfRange={outOfRange}
-      fee={data.fee}
-      protocol={data.protocol}
-      isStaked={data.isStaked}
-      tokenId={data.tokenId}
-      detailMode={detailMode}
-    />
-  )
-})
-
-const DetailInfoTitle = styled.div<{ $isMobile?: boolean }>`
-  display: flex;
-  gap: 8px;
-  font-size: 16px;
-  font-weight: 600;
-  flex-direction: ${({ $isMobile }) => ($isMobile ? 'column' : 'row')};
-`
-
-const DetailInfoDesc = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  font-size: 12px;
-  font-weight: 400;
-`
-
-const DetailInfoLabel = styled(Text)`
-  color: ${({ theme }) => theme.colors.textSubtle};
-  font-weight: 600;
-  font-size: 12px;
-`
-
-interface IPositionItemDetailProps {
-  chainId: number
-  currency0?: Currency
-  currency1?: Currency
-  removed: boolean
-  outOfRange: boolean
-  desc?: React.ReactNode
-  link?: string
-  tokenId?: bigint
-  fee: number
-  isStaked?: boolean
-  protocol: Protocol
-  totalPriceUSD: number
-  amount0?: CurrencyAmount<Token>
-  amount1?: CurrencyAmount<Token>
-  pool?: PoolInfo | null
-  detailMode?: boolean
-}
-
-export const PositionItemSkeleton = () => {
-  return (
-    <Container>
-      <Skeleton width={48} height={48} variant="circle" />
-      <div>
-        <Skeleton width={40} height={10} mb="4px" />
-        <Skeleton width={60} height={24} />
-      </div>
-    </Container>
-  )
-}
-
-const PositionDetailInfo = memo(
-  ({
-    currency0,
-    currency1,
-    removed,
-    outOfRange,
-    desc,
-    tokenId,
-    fee,
-    isStaked,
-    protocol,
-    totalPriceUSD,
-    amount0,
-    amount1,
-    pool,
-  }: IPositionItemDetailProps) => {
-    const { t } = useTranslation()
-    const { theme } = useTheme()
-    const { isMobile, isTablet } = useMatchBreakpoints()
-
-    const cakePrice = useCakePrice()
-    const stackedTokenId = useMemo(() => (tokenId ? [tokenId] : []), [tokenId])
-    // @todo @ChefJerry this only support v3
-    const {
-      tokenIdResults: [pendingCake],
-    } = useStakedPositionsByUser(stackedTokenId)
-    const earningsAmount = useMemo(() => +formatBigInt(pendingCake || 0n, 4), [pendingCake])
-    const earningsBusd = useMemo(() => {
-      return new BigNumber(earningsAmount).times(cakePrice.toString()).toNumber()
-    }, [cakePrice, earningsAmount])
-
-    const title = useMemo(
-      () =>
-        isTablet || isMobile ? (
-          <DetailInfoTitle $isMobile>
-            <Row>
-              <DoubleCurrencyLogo
-                size={24}
-                currency0={currency0}
-                currency1={currency1}
-                showChainLogo
-                innerMargin="-10px"
-              />
-              <FeeTier type={protocol} fee={fee} />
-            </Row>
-            <Row>
-              <Text bold>{`${currency0?.symbol} / ${currency1?.symbol} LP`}</Text>
-              {tokenId ? <Text color="textSubtle">(#{tokenId.toString()})</Text> : null}
-            </Row>
-            <Row>
-              {isStaked && (
-                <Tag variant="primary60" mr="8px">
-                  {t('Farming')}
-                </Tag>
-              )}
-              {protocol === Protocol.V3 && <RangeTag lowContrast removed={removed} outOfRange={outOfRange} />}
-            </Row>
-          </DetailInfoTitle>
-        ) : (
-          <DetailInfoTitle>
-            <Text bold>{`${currency0?.symbol} / ${currency1?.symbol} LP`}</Text>
-            {tokenId ? <Text color="textSubtle">(#{tokenId.toString()})</Text> : null}
-            <FeeTier type={protocol} fee={fee} />
-            <TagCell>
-              {isStaked && (
-                <Tag variant="primary60" mr="8px">
-                  {t('Farming')}
-                </Tag>
-              )}
-              {protocol === Protocol.V3 && <RangeTag lowContrast removed={removed} outOfRange={outOfRange} />}
-            </TagCell>
-          </DetailInfoTitle>
-        ),
-      [currency0, currency1, fee, isMobile, isTablet, isStaked, outOfRange, protocol, removed, t, tokenId],
-    )
-
-    return (
-      <>
-        {title}
-        <DetailInfoDesc>
-          <Row>{desc}</Row>
-          <Row gap="sm">
-            <FiatNumberDisplay
-              prefix="~"
-              value={totalPriceUSD}
-              style={{ color: theme.colors.textSubtle, fontSize: '12px' }}
-              showFullDigitsTooltip={false}
-            />
-            ({amount0 ? amount0.toFixed(6) : 0} {currency0?.symbol ?? '-'} / {amount1 ? amount1.toFixed(6) : 0}{' '}
-            {currency1?.symbol ?? '-'})
-          </Row>
-          <Row gap="8px">
-            <DetailInfoLabel>APR: </DetailInfoLabel>
-            {pool ? <PoolApyButton pool={pool} /> : <Skeleton width={60} />}
-          </Row>
-          {earningsAmount > 0 && (
-            <Row gap="8px">
-              <DetailInfoLabel>
-                {t('CAKE earned')}: {earningsAmount} (~${earningsBusd})
-              </DetailInfoLabel>
-            </Row>
-          )}
-        </DetailInfoDesc>
-      </>
-    )
-  },
-)
-
-const V3UnstakeModalContent: React.FC<IPositionItemDetailProps> = (props) => {
-  const { chainId, link, outOfRange, isStaked } = props
-  const { t } = useTranslation()
-  const linkWithChain = useMemo(
-    () => (link ? addQueryToPath(link, { chain: chainNames[chainId] }) : link),
-    [link, chainId],
-  )
-  return (
-    <>
-      <PositionDetailInfo {...props} />
-      {linkWithChain ? (
-        <Box mt="8px">
-          <NextLink href={linkWithChain}>
-            {outOfRange && !isStaked ? (
-              <Button
-                external
-                variant="primary"
-                width="100%"
-                as="a"
-                href={linkWithChain}
-                style={{ whiteSpace: 'nowrap' }}
-              >
-                {t('View Position')}
-              </Button>
-            ) : (
-              <Button variant="tertiary" width="100%" as="a">
-                {t('Manage Position')}
-              </Button>
-            )}
-          </NextLink>
-        </Box>
-      ) : null}
-    </>
-  )
-}
-
-export const PositionItemDetail = (props: IPositionItemDetailProps) => {
-  const { link, currency0, currency1, removed, outOfRange, tokenId, isStaked, detailMode, chainId } = props
-
-  const { isDesktop } = useMatchBreakpoints()
-  const linkWithChain = useMemo(
-    () => (link ? addQueryToPath(link, { chain: chainNames[chainId] }) : link),
-    [link, chainId],
-  )
-
-  if (!(currency0 && currency1)) {
-    return <PositionItemSkeleton />
-  }
-
-  const content = (
-    <Container>
-      {isDesktop && (
-        <TokenPairImage
-          width={48}
-          height={48}
-          variant="inverted"
-          primaryToken={currency0}
-          secondaryToken={currency1.wrapped}
-          withChainLogo
-        />
-      )}
-      <DetailsContainer>
-        <Column gap="8px">
-          <PositionDetailInfo {...props} />
-        </Column>
-        <Column justifyContent="flex-end">
-          <ActionPanel
-            currency0={currency0}
-            currency1={currency1}
-            isStaked={isStaked}
-            protocol={props.protocol}
-            fee={props.fee}
-            removed={removed}
-            tokenId={tokenId}
-            outOfRange={outOfRange}
-            modalContent={<V3UnstakeModalContent {...props} />}
-            detailMode={detailMode}
-          />
-        </Column>
-      </DetailsContainer>
-    </Container>
-  )
-
-  if (!linkWithChain) {
-    return content
-  }
-  return <StyledNextLink href={linkWithChain}>{content}</StyledNextLink>
-}
-
-const ActionPanelContainer = styled(Flex)`
-  flex-direction: row;
-  gap: 8px;
-  height: 48px;
-
-  & button {
-    flex: 1;
-  }
-`
-
-interface IActionPanelProps {
-  removed: boolean
-  outOfRange: boolean
-  tokenId?: bigint
-  isStaked?: boolean
-  detailMode?: boolean
-  modalContent: React.ReactNode
-  currency0: Currency
-  currency1: Currency
-  protocol: Protocol
-  fee: number
-}
-const ActionPanel = ({
-  currency0,
-  currency1,
-  isStaked,
-  removed,
-  outOfRange,
-  tokenId,
-  modalContent,
-  detailMode,
-  protocol,
-  fee,
-}: IActionPanelProps) => {
-  const { t } = useTranslation()
-  const { onStake, onUnstake, onHarvest, attemptingTxn } = useFarmV3Actions({
-    tokenId: tokenId?.toString() ?? '',
-    onDone: () => {},
-  })
-  const stakeModal = useModalV2()
-  const { switchNetworkIfNecessary, isLoading: isSwitchingNetwork } = useCheckShouldSwitchNetwork()
-
-  const handleStakeAndCheckInactive = useCallback(async () => {
-    logGTMClickStakeFarmEvent()
-    if (outOfRange && !isStaked) {
-      stakeModal.onOpen()
-    } else {
-      await switchNetworkIfNecessary(currency0.chainId)
-      await onStake()
-    }
-  }, [isStaked, onStake, outOfRange, stakeModal, switchNetworkIfNecessary, currency0.chainId])
-
-  const handleStake = useCallback(async () => {
-    logGTMClickStakeFarmEvent()
-    await switchNetworkIfNecessary(currency0.chainId)
-    await onStake()
-  }, [onStake, switchNetworkIfNecessary, currency0.chainId])
-
-  const handleUnStake = useCallback(async () => {
-    await switchNetworkIfNecessary(currency0.chainId)
-    await onUnstake()
-  }, [onUnstake, switchNetworkIfNecessary, currency0.chainId])
-
-  const handleHarvest = useCallback(async () => {
-    await switchNetworkIfNecessary(currency0.chainId)
-    await onHarvest()
-  }, [onHarvest, switchNetworkIfNecessary, currency0.chainId])
-
-  const preventDefault = useCallback((e: React.MouseEvent) => {
-    e.preventDefault()
-  }, [])
-
-  const handleIncreasePosition = useCallback(() => {
-    if (protocol === Protocol.V3) {
-      window.open(
-        `/increase/${currency0.wrapped.address}/${currency1.wrapped.address}/${fee}/${tokenId}`,
-        '_blank',
-        'noopener',
-      )
-    }
-  }, [currency0.wrapped.address, currency1.wrapped.address, fee, protocol, tokenId])
-
-  const handleDecreasePosition = useCallback(() => {
-    if (protocol === Protocol.V3) {
-      window.open(
-        `/decrease/${currency0.wrapped.address}/${currency1.wrapped.address}/${fee}/${tokenId}`,
-        '_blank',
-        'noopener',
-      )
-    }
-  }, [currency0.wrapped.address, currency1.wrapped.address, fee, protocol, tokenId])
-
-  const stakeButton = useMemo(
-    () => (
-      <>
-        <Button
-          scale="md"
-          width={['100px']}
-          style={{ alignSelf: 'center' }}
-          onClick={handleStakeAndCheckInactive}
-          disabled={attemptingTxn || isSwitchingNetwork}
-        >
-          {t('Stake')}
-        </Button>
-        <StakeModal
-          isOpen={stakeModal.isOpen}
-          staking={outOfRange && !isStaked}
-          onStake={handleStake}
-          onDismiss={stakeModal.onDismiss}
-        >
-          {modalContent}
-        </StakeModal>
-      </>
-    ),
-    [
-      isSwitchingNetwork,
-      handleStake,
-      handleStakeAndCheckInactive,
-      isStaked,
-      modalContent,
-      t,
-      outOfRange,
-      stakeModal,
-      attemptingTxn,
-    ],
-  )
-
-  const unstakeButton = useMemo(
-    () => (
-      <>
-        <Button
-          scale="md"
-          width={['100px']}
-          style={{ alignSelf: 'center' }}
-          variant="secondary"
-          onClick={stakeModal.onOpen}
-          disabled={attemptingTxn || isSwitchingNetwork}
-        >
-          {t('Unstake')}
-        </Button>
-        <StakeModal
-          isOpen={stakeModal.isOpen}
-          staking={outOfRange && !isStaked}
-          onUnStake={handleUnStake}
-          onDismiss={stakeModal.onDismiss}
-        >
-          {modalContent}
-        </StakeModal>
-      </>
-    ),
-    [isSwitchingNetwork, handleUnStake, isStaked, modalContent, t, outOfRange, stakeModal, attemptingTxn],
-  )
-  const showLpChangeActions = useMemo(() => {
-    if (protocol === Protocol.V3 && isStaked) return false
-    return true
-  }, [isStaked, protocol])
-
-  if (detailMode) {
-    return (
-      <ActionPanelContainer onClick={preventDefault}>
-        {showLpChangeActions ? (
-          <>
-            {!removed && (
-              <IconButton mr="6px" variant="secondary" onClick={handleDecreasePosition}>
-                <MinusIcon color="primary" width="14px" />
-              </IconButton>
-            )}
-            <IconButton variant="secondary" onClick={handleIncreasePosition}>
-              <AddIcon color="primary" width="14px" />
-            </IconButton>
-          </>
-        ) : null}
-        {isStaked ? unstakeButton : !removed ? stakeButton : null}
-        {isStaked && !removed ? (
-          <Button width={['100px']} scale="md" disabled={attemptingTxn || isSwitchingNetwork} onClick={handleHarvest}>
-            {attemptingTxn ? t('Harvesting') : t('Harvest')}
-          </Button>
-        ) : null}
-      </ActionPanelContainer>
-    )
-  }
-  return (
-    <ActionPanelContainer onClick={preventDefault}>
-      {!isStaked && !removed ? stakeButton : null}
-      {isStaked && !removed ? (
-        <Button width={['100px']} scale="md" disabled={attemptingTxn || isSwitchingNetwork} onClick={handleHarvest}>
-          {attemptingTxn ? t('Harvesting') : t('Harvest')}
-        </Button>
-      ) : null}
-    </ActionPanelContainer>
-  )
-}
