+++ apps/web/src/views/universalFarms/components/index.tsx
@@ -1,7 +1,7 @@
-export * from './PoolsFilterPanel'
-export * from './PoolsBanner'
+export * from './PoolListItemAction'
 export * from './PoolListView'
+export * from './PoolsBanner'
+export * from './PoolsFilterPanel'
+export * from './PositionItem'
 export * from './StyledCard'
 export * from './useColumnConfig'
-export * from './PoolListItemAction'
-export * from './PositionItem'
